move bigbang/domain to bigbang/analysis/utils; generals get_froms to other personal headers, notebook update
