+++ bigbang/analysis/utils.py
@@ -1,4 +1,5 @@
 import datetime
+import math
 import os
 import re
 import logging
@@ -130,3 +131,74 @@ def clean_datetime(df: pd.DataFrame) -> pd.DataFrame:
         for dt in df.loc[index, "date"].values
     ]
     return df
+
+
+email_regex = r"[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+"
+domain_regex = r"[@]([a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+)$"
+
+
+def extract_email(from_field):
+    """
+    Returns an email address from a string.
+    """
+    match = re.search(email_regex, from_field)
+
+    if match is not None:
+        return match[0].lower()
+
+    else:
+        return None
+
+
+def extract_domain(from_field):
+    """
+    Returns the domain of an email address from a string.
+    """
+    match = re.search(email_regex, from_field)
+
+    if match is not None:
+        return re.search(domain_regex, match[0])[1]
+
+    else:
+        return None
+
+
+def domain_entropy(domain, froms):
+    """
+    Compute the entropy of the distribution of counts of email prefixes
+    within the given archive.
+
+    Parameters
+    ---------------
+
+    domain: string
+        An email domain
+
+    froms: pandas.DataFrame
+        A pandas.DataFrame with From fields, email address, and domains.
+        See the Archive method ``get_froms()``
+
+
+    Returns
+    --------
+
+    entropy: float
+    """
+
+    domain_messages = froms[froms["domain"] == domain]
+
+    n_D = domain_messages.shape[0]
+
+    entropy = 0
+
+    emails = domain_messages["email"].unique()
+
+    for em in emails:
+        em_messages = domain_messages[domain_messages["email"] == em]
+        n_e = em_messages.shape[0]
+
+        p_em = float(n_e) / n_D
+
+        entropy = entropy - p_em * math.log(p_em)
+
+    return entropy
