+++ bigbang/archive.py
@@ -14,7 +14,7 @@ import pandas as pd
 from pathlib import Path
 import pytz
 
-import bigbang.domain as domain
+import bigbang.analysis.utils as analysis_utils
 import bigbang.mailman as mailman
 from bigbang.parse import get_date, get_text
 import bigbang.process as process
@@ -50,7 +50,7 @@ class Archive(object):
     activity = None
     threads = None
     entities = None
-    froms = None
+    preprocessed = None
 
     def __init__(self, data, archive_dir=CONFIG.mail_path, mbox=False):
         """
@@ -122,6 +122,7 @@ class Archive(object):
 
         # convert any null fields to None -- csv saves these as nan sometimes
         self.data = self.data.where(pd.notnull(self.data), None)
+        self.preprocessed = {}
 
         try:
             # set the index to be the Message-ID column
@@ -250,33 +251,39 @@ class Archive(object):
 
         return activity
 
-    def get_froms(self):
+    def get_personal_headers(self, header="From"):
         """
         Returns a dataframe with a row for every message of the archive, containing
         column entries for:
 
-         - The From field of the email
+         - The personal header specified. Defaults to "From". Could be "Repy-To".
          - The email address extracted from the From field
          - The domain of the From field
 
         This dataframe is computed the first time this method is called and then cached.
 
+
+        Parameters
+        ------------
+
+        header: string, default "From"
+
         Returns
         ----------
 
-        froms: pandas.DataFrame
+        data: pandas.DataFrame
         """
-        if self.froms is not None:
-            return self.froms
+        if header in self.preprocessed:
+            return self.preprocessed[header]
         else:
-            emails = self.data["From"].apply(domain.extract_email)
-            domains = self.data["From"].apply(domain.extract_domain)
-            self.froms = pd.concat(
-                [self.data["From"], emails, domains],
+            emails = self.data[header].apply(analysis_utils.extract_email)
+            domains = self.data[header].apply(analysis_utils.extract_domain)
+            self.preprocessed[header] = pd.concat(
+                [self.data[header], emails, domains],
                 axis=1,
-                keys=["From", "email", "domain"],
+                keys=[header, "email", "domain"],
             )
-            return self.froms
+            return self.preprocessed[header]
 
     def get_threads(self, verbose=False):
         """Get threads."""
